Remove unused LegacyCardAccountRangeRepository
